pyproject.toml
[build-system]
requires = ["poetry-core>=1.2.2"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
python_version = 3.10
show_error_context = true
pretty = true
show_column_numbers = true
warn_return_any = true
warn_unused_configs = true
show_error_codes = true
exclude = ['docs/']

[tool.poetry]
name = "volttron-platform-driver"
version = "2.0.0rc0"
description = "The Platform Driver agent is a special purpose agent that manages communication between the Volttron platform and devices. The Platform driver features a number of endpoints for collecting data and sending control signals using the message bus and automatically publishes data to the bus on a specified interval."
authors = ["VOLTTRON Team <volttron@pnnl.gov>"]
license = "Apache License 2.0"
readme = "README.md"
repository = "https://github.com/eclipse-volttron/platform-driver-agent"
homepage = "https://github.com/eclipse-volttron/platform-driver-agent"
keywords = []
packages = [ { include = "platform_driver", from = "src" } ]
classifiers = [
    "Programming Language :: Python :: 3 :: Only",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Information Technology",
    "Intended Audience :: Developers",
    "Intended Audience :: Other Audience",
    "License :: OSI Approved :: Apache Software License"
]
include = ["volttron-platform-driver-*-default-vip-id"]

[tool.poetry.dependencies]
pydantic = ">=2"
python = ">=3.10,<4.0"
treelib = ">=1.7.0"
volttron-core = ">=2.0.0rc0"
volttron-lib-base-driver = ">=2.0.0rc0"
# volttron-lib-topic-tree = { path = "../volttron-lib-topic-tree", develop = true }


[tool.poetry.group.dev.dependencies]
pytest = ">=6.2.5,<7.0.0"
pytest-cov = ">=3.0.0,<4.0.0"
mock = ">=4.0.3,<5.0.0"
pre-commit = ">=2.17.0,<3.0.0"
yapf = ">=0.32.0,<1.0.0"
toml = ">=0.10.2,<1.0.0"
isort = ">=5.10.1,<6.0.0"
safety = ">=1.10.3,<2.0.0"
mypy = ">=0.942,<1.0"
coverage = ">=6.3.2,<7.0.0"


[tool.poetry.group.documentation.dependencies]
Sphinx = ">=4.5.0,<5.0.0"
sphinx-rtd-theme = ">=1.0.0,<2.0.0"

[tool.poetry.scripts]
volttron-platform-driver = "platform_driver.agent:main"

[tool.yapf]
based_on_style = "pep8"
spaces_before_comment = 4
column_limit = 99
split_before_logical_operator = true

[tool.yapfignore]
ignore_patterns = [
    ".venv/**",
    ".pytest_cache/**",
    "dist/**",
    "docs/**"
]
